+++ edzed/utils/tconst.py
@@ -6,6 +6,6 @@ SEC_PER_DAY = 86_400
 SEC_PER_HOUR = 3_600
 SEC_PER_MIN = 60
 MONTH_NAMES = (
-    None,
+    '???',
     'January', 'February', 'March', 'April', 'May', 'June',
     'July', 'August', 'September', 'October', 'November', 'December')