+++ tests/test_outputfunc1.py
@@ -1,5 +1,6 @@
 """
 Test the OutputFunc block.
+Part 1/2 - async
 """
 
 # pylint: disable=missing-docstring, no-self-use, protected-access
@@ -17,15 +18,6 @@ from .utils import *
 
 pytest_plugins = ('pytest_asyncio',)
 pytestmark = pytest.mark.asyncio
-
-
-async def test_argument_checks(circuit):
-    """Test f_args and f_kwargs validation."""
-    for val in (None, 0, "string", edzed.UNDEF, {1:2}, {'xy'}, ["A", "B", 1], (True, False)):
-        with pytest.raises(TypeError, match="f_args"):
-            edzed.OutputFunc('err', func=lambda x: 0, on_error=None, f_args=val)
-        with pytest.raises(TypeError, match="f_kwargs"):
-            edzed.OutputFunc('err', func=lambda x: 0, on_error=None, f_kwargs=val)
 
 
 async def output_func(circuit, *, log, v2=2, on_error=None, mstop=True, **kwargs):